Move createServer under src/shared
