+++ benchmarks/postgres/ingestion.benchmark.ts
@@ -4,8 +4,8 @@ import os from 'os'
 import { performance } from 'perf_hooks'
 
 import { IEvent } from '../../src/idl/protos'
-import { createServer } from '../../src/server'
 import { EventsProcessor } from '../../src/shared/ingestion/process-event'
+import { createServer } from '../../src/shared/server'
 import { UUIDT } from '../../src/shared/utils'
 import { LogLevel, PluginsServer, SessionRecordingEvent, Team } from '../../src/types'
 import { getFirstTeam, resetTestDatabase } from '../../tests/helpers/sql'
