+++ benchmarks/vm/memory.benchmark.ts
@@ -1,6 +1,6 @@
 import { PluginEvent } from '@posthog/plugin-scaffold/src/types'
 
-import { createServer } from '../../src/server'
+import { createServer } from '../../src/shared/server'
 import { Plugin, PluginConfig, PluginConfigVMReponse } from '../../src/types'
 import { createPluginConfigVM } from '../../src/worker/vm/vm'
 import { commonOrganizationId } from '../../tests/helpers/plugins'
