+++ src/main/pluginsServer.ts
@@ -4,8 +4,8 @@ import { FastifyInstance } from 'fastify'
 import Redis from 'ioredis'
 import * as schedule from 'node-schedule'
 
-import { createServer } from '../server'
 import { defaultConfig } from '../shared/config'
+import { createServer } from '../shared/server'
 import { status } from '../shared/status'
 import { createRedis, delay } from '../shared/utils'
 import { PluginsServer, PluginsServerConfig, Queue, ScheduleControl } from '../types'
