+++ tests/postgres/vm.timeout.test.ts
@@ -1,4 +1,4 @@
-import { createServer } from '../../src/server'
+import { createServer } from '../../src/shared/server'
 import { PluginsServer } from '../../src/types'
 import { createPluginConfigVM } from '../../src/worker/vm/vm'
 import { pluginConfig39 } from '../helpers/plugins'
